refactor: extract _ActiveContext to break context cycle
